Add update.yml. Runs nightly to check for new commits.
If new commits are found, the target branch on forks is synced with upstream, and build_iAPS.yml is launched. Updates will hence be made available in TestFlight within less than a day. Users will have to decide if they want to enable Automatic Updates in the  TestFlight app or not.

Scheduled workflows in GitHub actions will be cancelled in the event of 60 days of no repository activity. A "keep alive" action will generate empty commits on upstream if there is no activity after 50 days to avoid inactivation of the scheduled workflow. This commit will be picked up by update.yml on forks the next day, so that there is repository activity also on forks.

The reason for running the keepalive action on upstream only, is that upstream and forks should then be on the same commit ref. This would not always be the case if the keepalive action ran on the forks instead. But then the scheduled workflow must be enabled in the upstream repo (Artificial-Pancreas/iAPS).

Changes to build_iAPS.yml:
- Added "on: workflow_call" to allow it to be launched by update.yml
- Removed the "secrets" job", this can still be run from the identifiers and certificates workflows to validate repository secrets.
- Removed checkout option "submodules: recursive"