Improve the bolus calculator. (#98)
When entering carbs and coming to the bolus recommendation screen, use the EventualBG instead of the minPredBG when computing recommended bolus. Only used when directly combined with carbs. If not everything else is as before.

To not make things overly complicated I'm using this new bolus recommendation for Watch App also, but always, until I find a neat way of collecting the manual state from the BolusStateModel.

For testing. Tested for about a week without any dramatic adverse effects, but needs more testing from more users.