Add a 'Push Settings' button to the Nightscount Settings screen that does a force upload of preferences / statistics
Moved the check for "later then 20:00" out of statistics(), and gave
a parameter to statistics() to "force" the run, overriding the "once per
day" check ...

tested on my phone, can press the button repeatedly and it is updating,
no crashes, and even tested going into preferences, changing my maxIOB and
force the upload, works like like a charm.

the only thing I don't like, and you may have a clean idea on how to do,
is that when you press 'Push Settings', it doesn't *appear* to do anything,
even though I know it does.

but it does do as expected / desired.