Extended pushSettings to include Profile upload
Added code to upload{Profile,Preferences,Statistics} to flag when
upload failed or not, and extended that into the button so that
not only does it determine if it worked, but let's you know which
of the uploads didn't work.

It appears to work fine, I am getting intermediant failed states for
the uploads, but a second push shows them all as OK, as if it was a
temporary network issue.