New UI/UX
New UI/UX

Header:
Make the current glucose more prominent and taking up more space.
Less strings and more graphics. Replace the boring and cluttered strings with dynamic graphics.

Test tubes in frosted glass for IOB and COB, displaying amount as coloured liquid.

Pump status using dynamic graphics instead of strings. Both battery icon and reservoir vial image is now illustrated with a dynamic fill.

Loop button is now more discrete. Normally iAPS is functioning well and you shouldn’t need to worry about if looping or not, because it will be looping. However, when not looping the loop button the button will appear with red text.

When tapping the loop button you will now see all info. Both chart, header and the openAPS suggestion. This pop-up info is now toggled with the loop button, meaning if you tap button again the pop-up disappears (or if tapping inside of pop-up as before).

Chart and Info panel

The chart and info panel is now less cluttered.
Eventual glucose is displayed more readable. The legends are removed and only displayed when tapped once. Tapping once again will remove them (thanks for this idea @nas10 ). To indicate that there in fact are legends available small coloured dots are added.

Bolus Progress is now displayed as layer on top of chart. Bigger and with a clearer stop button.

When legends are displayed they will now appear in the future x-side:

When Max IOB setting is 0 this will now be displayed more prominently:

Profiles are now displayed in chart. Duration and target glucose will displayed similar to the temp targets.
B
Profile name will be displayed centred in info panel, as displayed below.

Button Panel

The extra space from the previous profile button is now used for a bigger header instead.

New profile button. When profile selected, this button will be highlighted, similar to the Loop app. A enabled profile will be deactivated with just tap, also similar to the Loop app (thank you Loop authors for the idea (didn’t borrow any code)).

New history button. Instead of tapping the chart you now have a button to access the history.

Vertical Scroll View and Statistics Preview

To make the button panel and home view both less cluttered and more extendable I’ve removed the statistics button and introduced a vertical scroll.

To see a preview of statistics just scroll down. To see full statistics view tap the preview (now just TIR for current day is displayed, but more info will be added later).

The idea with the vertical scroll is also to make the home more more extendable with future updates. Later we can add for instance an Insulin chart here, or a meal chart, or whatever really. Any info you would like to sometimes look at, but not always need to see.

This info is only displayed when scrolling down, meaning it will not add any clutter or extra ”info overload” when using the iAPS app as you’re used to do.

Temp Targets
Temp targets will still work as before. For those of you who really need to use both TTs and profiles simultaneously you can display the Temp Target button in UI/UX settings. Default is to not display this button. I recommend not normally just use one of these, but I can understand the need for sometimes having to use both, which is why I decided to not remove this option entirely.

Tested a couple of weeks on my own phone and with different simulators.
To do: When activating a new profile the new highlight and the name in info panel will appear instantly, but there is a delay before the new current profile is illustrated also in the chart. I'll fix this ASAP.
To do: When activating a new profile the new highlight and the name in info panel will appear instantly, but there is a delay before the new current profile is illustrated also in the chart. I'll fix this ASAP.
clean

(cherry picked from commit 62af4eaca29ffa4103d26075a0476d54b1aacc1c)

Fix bad merge