Once tested, I will extend the docs to add a section to enable.
Its essentially the same instructions as build_iAPS, but until
this is in dev, the code is hard coded so that it can run out
a separate default branch for testing.

Once tested and in dev, I will change the references to dev
to be ${{ github.ref_name }}, and add docs.

Basically, I don't want to advertise until we've confirmed
it, *but*, it has to be in dev for us to properly test it ...

sort of a chicken and egg situation.