Allow for per branch auto build
Triggered by a push ( ie. sync with master ), this code will check
to see if:

  - a variable exists called AUTO_BUILD_BRANCHES exists
  - said variable contains the branch

For example, if this workflow exists on the main branch, and
AUTO_BUILD_BRANCHES=main, the app will be auto-built and uploaded
to Testflight ...

If AUTO_BUILD_BRANCHES=dev though, the workflow will still be
triggered, but stop right at the start, at the Validate stage,
since the auto_build_check will return false.

This does not affect manual builds ... build process *if* the
branch is in the AUTO_BUILD_BRANCHES list *or* it a manual build.

Have tested as:

  - a manual build
  - a push where AUTO_BUILD_BRANCHES contains the branch
  - a push where AUTO_BUILD_BRANCHES doesn't contain the branch