CGM time remaining
displayed in header View. Display both the time remaining and (illustrate) the sensor age.

When using xdrip or glucoseDirect you will have the option to enter sensor life-span in CGM settings. This setting serves two objectives: indirectly tell iAPS the type of sensor and allow for a correct display of an eventual extended sensor life-span .